change data access strategy
