+++ app/src/main/res/values/strings.xml
@@ -12,7 +12,7 @@
     <string name="Ufa">Уфа</string>
     <string name="Tula">Тула</string>
     <string name="Anapa">Анапа</string>
-    <string name="network_error">Ошибка подключения ,Загруженных жанных нет.</string>
+    <string name="network_error">Ошибка подключения,Загруженных данных нет.</string>
     <string name="network_error_load">Ошибка подключения,Показаны загруженные данные.</string>
     <string name="error">Неизвестная ошибка.</string>
     <string name="textview">TextView</string>
